# clients/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "node-flow",
    "version": "1",
    "description": "Workflow orchestration. Definitions are declarative JSON; workers are language-agnostic and lease tasks from a queue."
  },
  "components": {
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Long-lived, revocable, scoped. For CLIs and CI."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived token from POST /v1/auth/token. For worker fleets: a captured token is useless within the hour."
      }
    }
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/v1/auth/api-keys": {
      "get": {
        "operationId": "auth_listApiKeys",
        "summary": "GET /v1/auth/api-keys",
        "tags": [
          "auth"
        ],
        "x-required-scopes": [
          "admin"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      },
      "post": {
        "operationId": "auth_createApiKey",
        "summary": "Create an API key",
        "description": "The token is returned once and cannot be recovered.",
        "tags": [
          "auth"
        ],
        "x-required-scopes": [
          "admin"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "scopes": {
                    "default": [],
                    "maxItems": 64,
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/auth/api-keys/{id}": {
      "delete": {
        "operationId": "auth_revokeApiKey",
        "summary": "DELETE /v1/auth/api-keys/{id}",
        "tags": [
          "auth"
        ],
        "x-required-scopes": [
          "admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "auth_me",
        "summary": "Who the caller is",
        "description": "Namespace-free on purpose: a browser holding a session cookie knows neither its namespace slug nor its scopes until it asks, and every other route needs the namespace in the path.",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
...
    "/v1/auth/service-accounts": {
      "get": {
        "operationId": "auth_listServiceAccounts",
        "summary": "GET /v1/auth/service-accounts",
        "tags": [
          "auth"
        ],
        "x-required-scopes": [
          "admin"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      },
      "post": {
        "operationId": "auth_createServiceAccount",
        "summary": "Create a service account",
        "description": "The secret is returned once and cannot be recovered.",
        "tags": [
          "auth"
        ],
        "x-required-scopes": [
          "admin"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "scopes": {
                    "default": [],
                    "maxItems": 64,
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/auth/service-accounts/{id}": {
      "delete": {
        "operationId": "auth_disableServiceAccount",
        "summary": "DELETE /v1/auth/service-accounts/{id}",
        "tags": [
          "auth"
        ],
        "x-required-scopes": [
          "admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
...
    "/v1/auth/token": {
      "post": {
        "operationId": "auth_token",
        "summary": "Exchange a service-account key and secret for a short-lived token",
        "tags": [
          "auth"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "keyId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "secret": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "keyId",
                  "secret"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/auth/whoami": {
      "get": {
        "operationId": "auth_whoami",
        "summary": "The caller as the server sees them, including scopes",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
...
  }
}